Modified scripts to work with paths continaing spaces.

+++ AirshipLib/update_library_reference.sh
@@ -28,16 +28,18 @@ lib_base_name="$(echo $lib_name | awk -F '-' '{print $1}')"
 dest_lib_root="${SRCROOT}/../Airship"
 
 echo "remove old library $lib_base_name*.${EXECUTABLE_EXTENSION}"
-find "$dest_lib_root" -d 1 -name "$lib_base_name*.${EXECUTABLE_EXTENSION}" | xargs rm
+find "$dest_lib_root" -d 1 -name "$lib_base_name*.${EXECUTABLE_EXTENSION}" -exec rm {} \;
 echo "copy $lib_name from ${TARGET_BUILD_DIR} to $dest_lib_root"
 cp "${TARGET_BUILD_DIR}/$lib_name" "$dest_lib_root"
 
 
-for sample_prj_root in ${SRCROOT}/../*Sample
+for sample_prj_root in "${SRCROOT}/../*Sample"
 do
-    sample_prj_name=$(basename "$sample_prj_root")
+    sample_prj_name="$(basename "$sample_prj_root")"
     sample_prj_setting_file="$sample_prj_root/$sample_prj_name.xcodeproj/project.pbxproj"
     echo "update library reference in $sample_prj_setting_file"
-    sed "s/$lib_base_name[^ ]*\.${EXECUTABLE_EXTENSION}/$lib_name/g" "$sample_prj_setting_file">"/tmp/$sample_prj_name.tmp"
-    mv "/tmp/$sample_prj_name.tmp" "$sample_prj_setting_file"
+	if [[ -f "$sample_prj_setting_file" ]]; then
+    	sed "s/$lib_base_name[^ ]*\.${EXECUTABLE_EXTENSION}/$lib_name/g" "$sample_prj_setting_file" >"/tmp/$sample_prj_name.tmp"
+	    mv "/tmp/$sample_prj_name.tmp" "$sample_prj_setting_file"
+	fi
 done
